refactor: 역할을 모호하게 드러내는 config.ts / language.ts파일명을supportedLanguages.ts 와 noteLanguage.ts로 변경
