+++ src/features/editor/supportedLanguages.ts
@@ -1,7 +1,7 @@
 import {
   CODE_LANGUAGE_VALUES,
   type CodeLanguage,
-} from "@/lib/constants/languages";
+} from "@/lib/constants/noteLanguages";
 
 export const SUPPORTED_LANGUAGES = CODE_LANGUAGE_VALUES;
