+++ src/lib/constants/noteLanguages.ts
